test(wiki): split agent-scope CLI coverage
